+++ Logos/src/utopia/logos/database/access/many/url/link/DbLinks.scala
@@ -166,7 +166,7 @@ object DbLinks extends ManyLinksAccess with UnconditionalView
 			} }
 		}
 		else
-			Pair.twice(Vector.empty)
+			Pair.twice(Empty)
 	}
 	
 	private def ensureModelMaxLength(model: Model) = {