+++ Scribe/Scribe-Api/src/utopia/scribe/api/controller/logging/LogCleaner.scala
@@ -2,7 +2,7 @@ package utopia.scribe.api.controller.logging
 
 import utopia.flow.collection.immutable.range.Span
 import utopia.flow.collection.CollectionExtensions._
-import utopia.flow.collection.immutable.Pair
+import utopia.flow.collection.immutable.{Empty, Pair}
 import utopia.flow.generic.model.immutable.Model
 import utopia.flow.time.Now
 import utopia.flow.time.TimeExtensions._
@@ -127,7 +127,7 @@ object LogCleaner
 			// Uses the latest non-empty set of error messages
 			val errorMessages = ordered.reverseIterator
 				.findMap { o => NotEmpty(o.errorMessages) }
-				.getOrElse(Vector.empty)
+				.getOrElse(Empty)
 			// Uses only the latest non-empty set of details
 			val details = ordered.reverseIterator
 				.findMap { _.details.notEmpty }